agentos-docker/pyproject.toml
[project]
name = "agentos-docker"
version = "1.0.0"
requires-python = ">=3.12,<4"
readme = "README.md"
license-files = ["LICENSE"]
authors = [{ name = "Agno", email = "hello@agno.com" }]

dependencies = [
  # Agno
  "agno",
  "agno-infra",
  # API
  "fastapi[standard]",
  # Models
  "openai",
  "anthropic",
  # Database
  "pgvector",
  "psycopg[binary]",
  "sqlalchemy",
  "duckdb",
  # Tools
  "mcp",
  "reportlab",
  "duckduckgo-search",
  "python-dotenv",
  "python-pptx",
  # Tracing
  "opentelemetry-api",
  "opentelemetry-sdk",
  "openinference-instrumentation-agno",
]

[project.optional-dependencies]
dev = ["mypy", "ruff"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

[tool.ruff]
line-length = 120
exclude = [".venv*"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = [".venv*"]

[[tool.mypy.overrides]]
module = ["pgvector.*", "setuptools.*", "agno.*", "duckdb.*"]
ignore_missing_imports = true

[tool.uv.pip]
no-annotate = true
